pyproject.toml
[project]
name = "steganography"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1.0",
    "tk>=0.1.0",
    "pillow>=10.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
